[Decomp][Player] Cohesion split: extract movement-flag setters to PlayerMovement.cpp
