+++ src/game/Object/Player.cpp
@@ -3932,84 +3932,6 @@ void Player::DeleteOldCharacters(uint32 keepDays)
 }
 
 /**
- * @brief Forces or clears rooted movement for the player.
- *
- * @param enable True to root the player; false to unroot them.
- */
-void Player::SetRoot(bool enable)
-{
-    WorldPacket data;
-    BuildForceMoveRootPacket(&data, enable, 0);
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables water walking for the player.
- *
- * @param enable True to enable water walking; false to restore normal movement.
- */
-void Player::SetWaterWalk(bool enable)
-{
-    WorldPacket data;
-    BuildMoveWaterWalkPacket(&data, enable, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Placeholder for levitation support on this client version.
- *
- * @param enable Unused levitation state flag.
- */
-void Player::SetLevitate(bool enable)
-{
-    WorldPacket data;
-    BuildMoveLevitatePacket(&data, enable, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Enables or disables flying movement flags for the player.
- *
- * @param enable True to enable flight-related movement flags; false to clear them.
- */
-void Player::SetCanFly(bool enable)
-{
-    WorldPacket data;
-    BuildMoveSetCanFlyPacket(&data, enable, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Enables or disables feather fall movement for the player.
- *
- * @param enable True to enable feather fall; false to restore normal falling.
- */
-void Player::SetFeatherFall(bool enable)
-{
-    WorldPacket data;
-    BuildMoveFeatherFallPacket(&data, enable, 0);
-    SendMessageToSet(&data, true);
-
-    // start fall from current height
-    if (!enable)
-    {
-        SetFallInformation(0, GetPositionZ());
-    }
-}
-
-/**
- * @brief Enables or disables hover movement for the player.
- *
- * @param enable True to enable hovering; false to disable it.
- */
-void Player::SetHover(bool enable)
-{
-    WorldPacket data;
-    BuildMoveHoverPacket(&data, enable, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
  * @brief Moves the player to a new position and updates related state.
  *
  * @param x The destination X coordinate.
